requirements-travis.txt
Django
gunicorn
django-allauth
django-bootstrap4
google-auth
google-api-python-client
oauth2client
unittest2
crispy-forms
pillow
twilio
